+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingActivity.kt
@@ -8,9 +8,10 @@ import androidx.activity.viewModels
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import woowacourse.shopping.data.localdb.ShoppingDB
+import woowacourse.shopping.data.network.NetworkManager
 import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.data.repository.HttpProductRepository
-import woowacourse.shopping.data.repository.HttpProductServer
+import woowacourse.shopping.data.network.HttpProductServer
 import woowacourse.shopping.data.repository.RecentItemRepository
 import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.detail.DetailActivity
@@ -23,6 +24,7 @@ class ShoppingActivity : ComponentActivity() {
             productRepository = HttpProductRepository(HttpProductServer.baseUrl),
             cartRepository = CartRepository(database.cartItemDao()),
             recentItemRepository = RecentItemRepository(database.recentItemDao()),
+            networkManager = NetworkManager(applicationContext),
         )
     }
